Update Cavern museum
