+++ Team70_CavernPart/Assets/Art/Texture/Red/Red.mat
@@ -21,11 +21,11 @@ Material:
     serializedVersion: 3
     m_TexEnvs:
     - _BaseMap:
-        m_Texture: {fileID: 2800000, guid: db5dea22f0ef28a488b413facc5eaf02, type: 3}
+        m_Texture: {fileID: 2800000, guid: b65ea1158d1e6684e9e55a9bfc25283d, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _BumpMap:
-        m_Texture: {fileID: 2800000, guid: 22ce8838fbfb8e74d96993f1eebdf487, type: 3}
+        m_Texture: {fileID: 2800000, guid: 546c4eb3a8606e94e8d92f883a8fbda9, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _EmissionMap:
@@ -33,7 +33,7 @@ Material:
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MainTex:
-        m_Texture: {fileID: 2800000, guid: db5dea22f0ef28a488b413facc5eaf02, type: 3}
+        m_Texture: {fileID: 2800000, guid: b65ea1158d1e6684e9e55a9bfc25283d, type: 3}
         m_Scale: {x: 1, y: 1}
         m_Offset: {x: 0, y: 0}
     - _MetallicGlossMap:
@@ -71,7 +71,7 @@ Material:
     - _WorkflowMode: 1
     - _ZWrite: 1
     m_Colors:
-    - _BaseColor: {r: 0.8113208, g: 0.78160536, b: 0.76539695, a: 1}
+    - _BaseColor: {r: 0.8679245, g: 0.8504354, b: 0.8228907, a: 1}
     - _Color: {r: 1, g: 1, b: 1, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
     - _SpecColor: {r: 0.19999996, g: 0.19999996, b: 0.19999996, a: 1}
