--- Team70_CavernPart/Assets/Scripts/Util/GenericEventButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0d949abb08e233744b71542f31480dea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
